Move constants to constants.py
