+++ web3indexer/constants.py
@@ -1,0 +1,17 @@
+# keccak256("Transfer(address,address,uint256)")
+ERC721_TRANSFER_TOPIC = (
+    "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"
+)
+# keccak256("TransferSingle(address,address,address,uint256,uint256)")
+ERC1155_TRANSFER_SINGLE_TOPIC = (
+    "0xc3d58168c5ae7397731d063d5bbf3d657854427343f4c083240f7aacaa2d0f62"
+)
+# keccak256("TransferBatch(address,address,address,uint256[],uint256[])")
+ERC1155_TRANSFER_BATCH_TOPIC = (
+    "0x4a39dc06d4c0dbc64b70af90fd698a233a518aa5d07e595d983b8c0526c8f7fb"
+)
+ERC_165_IDENTIFIER = "01ffc9a7"
+ERC_721_IDENTIFIER = "80ac58cd"
+ERC_721_METADATA_IDENTIFIER = "5b5e139f"
+ERC_1155_IDENTIFIER = "d9b67a26"
+ERC_1155_METADATA_IDENTIFIER = "0e89341c"
